fix hostent struct for wii
h_addrtype and h_length are 16bit on wii network stack. closes #252

+++ external/bsd/include/netdb.h
@@ -84,8 +84,8 @@ extern __thread int h_errno;
 struct hostent {
 	char	*h_name;	/* official name of host */
 	char	**h_aliases;	/* alias list */
-	int	h_addrtype;	/* host address type */
-	int	h_length;	/* length of address */
+	short	h_addrtype;	/* host address type */
+	short	h_length;	/* length of address */
 	char	**h_addr_list;	/* list of addresses from name server */
 #define	h_addr	h_addr_list[0]	/* address, for backward compatibility */
 };
